New MinMax and AlphaBeta Added The gui is able to play to aiMoves and give a one move hint

+++ src/Client_Server/Huristics.java
@@ -3,5 +3,5 @@ package Client_Server;
 import Solver.Node;
 
 public interface Huristics {
-	int Evaluation(Node N);
+	double Evaluation(Node N);
 }
